drones that terminate early have NaN in state instead of zeros

+++ rotorpy/trajectories/minsnap.py
@@ -389,17 +389,17 @@ class BatchedMinSnap:
                     segment_idxs[i] = j
                     ts[i] = t_tmp[i] - self.t_keyframes[i,j]
                     break
-        segment_idxs = segment_idxs.to(self.device, non_blocking=True)
+        segment_idxs = segment_idxs.to(self.device, non_blocking=True).long()
         ts = ts.to(self.device, non_blocking=True)
 
 
         ts_expd = torch.cat([(ts**i).unsqueeze(-1) for i in reversed(range(0, 8))], dim=-1).unsqueeze(1)
         batch_idxs = torch.arange(self.num_trajs, device=self.device)
-        x_out = self.batch_polyval(self.xs[batch_idxs,segment_idxs], ts_expd).double()
-        x_dot_out = self.batch_polyval(self.x_dots[batch_idxs,segment_idxs], ts_expd[...,1:]).double()
-        x_ddot_out = self.batch_polyval(self.x_ddots[batch_idxs,segment_idxs], ts_expd[...,2:]).double()
-        yaw_out = self.batch_polyval(self.yaws[batch_idxs,segment_idxs], ts_expd).squeeze(-1).double()
-        yaw_dots_out = self.batch_polyval(self.yaw_dots[batch_idxs,segment_idxs], ts_expd[...,1:]).squeeze(-1).double()
+        x_out = self.batch_polyval(self.xs[batch_idxs.long(),segment_idxs], ts_expd).double()
+        x_dot_out = self.batch_polyval(self.x_dots[batch_idxs.long(),segment_idxs], ts_expd[...,1:]).double()
+        x_ddot_out = self.batch_polyval(self.x_ddots[batch_idxs.long(),segment_idxs], ts_expd[...,2:]).double()
+        yaw_out = self.batch_polyval(self.yaws[batch_idxs.long(),segment_idxs], ts_expd).squeeze(-1).double()
+        yaw_dots_out = self.batch_polyval(self.yaw_dots[batch_idxs.long(),segment_idxs], ts_expd[...,1:]).squeeze(-1).double()
 
         flat_output = {'x':x_out, 'x_dot':x_dot_out, 'x_ddot':x_ddot_out,
                        'yaw':yaw_out, 'yaw_dot':yaw_dots_out}
